SECURITY: Fix escaping in googleAnalyticsTrackUrl
This was failing to properly escape the values in the javascript
onclick event handler. It was also failing to validate the protocol
allowing javascript: urls

Last of all, it wasn't registering links in parser output, so it
could be used to bypass spam blacklists.

This change prevents relative urls from being used.

Bug: T333980
Change-Id: I95d0765651b8e96514e0e636abc32a667973e2b6